intial

+++ src/main/java/com/skyblockflipper/backend/api/RangePagination.java
@@ -11,11 +11,11 @@ public final class RangePagination {
     }
 
     public static Pageable pageable(Integer min, Integer max, int defaultSize, Sort sort) {
-        long safeDefaultSize = Math.max(1L, (long) defaultSize);
+        long safeDefaultSize = Math.max(1L, defaultSize);
         long safeMin = min == null ? 0L : Math.max(0L, min.longValue());
         long safeMax = max == null ? safeMin + safeDefaultSize - 1L : Math.max(safeMin, max.longValue());
         long size = safeMax - safeMin + 1L;
-        long clampedSizeLong = Math.max(1L, Math.min((long) MAX_PAGE_SIZE, size));
+        long clampedSizeLong = Math.max(1L, Math.min(MAX_PAGE_SIZE, size));
         int clampedSize = (int) clampedSizeLong;
         return OffsetLimitPageRequest.of(safeMin, clampedSize, sort);
     }
